Replace vagnrat with kind (#66)
* Aling with uostream and change image folder name

The ols scripts.vagrant/images is gone and replaced
by build/images. Adhere to that.

Signed-off-by: Nikolay Nikolaev <nnikolay@nnikolay-a01.vmware.com>

* Get rid of vagrant and default to kind

Upstream NSM now refers to Kind as the default. Make the
releaven build and documentation changes to reflect that in
the examples.

Signed-off-by: Nikolay Nikolaev <nnikolay@nnikolay-a01.vmware.com>

* Remove scripts/kind.sh

The latest Kind uses the K8s contexts to switch clusters.
Remove the unneeded script and update to docs on how
to use the contexts.

Signed-off-by: Nikolay Nikolaev <nnikolay@nnikolay-a01.vmware.com>