Update the proxy example to have three colored nginx instances
the proxy is now demonstrating how NSM uses labels and how to dynamically
take advantage of these. The Network Service Descriptor will also allow
for demoing the round robin between the endpoints, when there is no `color`
label, or the color does not exist in the set of labelled endpoints.

Signed-off-by: Nikolay Nikolaev <nnikolay@vmware.com>